Updates for MarketDataApp SDK compliance
- Added DotNetEnv .env loading via AddDotNetEnv().
- Replaced configuration keys with MARKETDATA_*.
- Use decimal type for all monetary model/request properties.
- Preserved decimal precision during JSON parsing.
- Added scalar overloads for endpoint methods.
- Updated tests, documentation, examples, CI, and integration configuration.
- Updated retry defaults and validation.
- Added/updated endpoint overload tests.

+++ src/MarketData/Funds/FundCandle.cs
@@ -12,7 +12,7 @@ namespace MarketData.Funds;
 /// <param name="Close">Closing NAV (wire field: <c>c</c>).</param>
 public record FundCandle(
     DateTimeOffset? Time,
-    double? Open,
-    double? High,
-    double? Low,
-    double? Close);
+    decimal? Open,
+    decimal? High,
+    decimal? Low,
+    decimal? Close);
